+++ src/shared/engine/HandleMovement.cpp
@@ -13,7 +13,7 @@ CommandTypeId HandleMovement::getTypeId() const {
 }
 
 bool HandleMovement::execute(state::Unit &unit, state::State& state, unsigned int new_x, unsigned int new_y) {
-    if( (unit.getPosition().getX()  + unit.getPosition().getY() - (new_x + new_y) ) < unit.getMovementRange() ){
+    if( (int)(unit.getPosition().getX()  + unit.getPosition().getY() - (new_x + new_y) ) < unit.getMovementRange() ){
         state::Position position(new_x,new_y);
         unit.setPosition(position);
     }
